requirements.txt
anaconda-client==1.2.2
appdirs==1.4.4
argh==0.26.2
asgiref==3.4.1
atomicwrites==1.4.0
backports.weakref==1.0.post1
bkcharts==0.2
black==19.10b0
boto==2.49.0
brotlipy==0.7.0
certifi==2020.12.5
comtypes==1.1.9
conda==4.3.16
cycler==0.10.0
cytoolz==0.11.0
dj-database-url==0.5.0
Django==3.2.9
django-heroku==0.3.1
entrypoints==0.3
et-xmlfile==1.0.1
fastcache==1.1.0
future==0.18.2
HeapDict==1.0.1
jdcal==1.4.1
json5==0.9.5
llvmlite==0.36.0
locket==0.2.1
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
mpmath==1.2.1
multipledispatch==0.6.0
mypy-extensions==0.4.3
olefile==0.46
parso==0.7.0
pathspec==0.7.0
patsy==0.5.1
pep8==1.7.1
pkginfo==1.7.0
ply==3.11
psycopg2==2.9.2
pycurl==7.43.0.6
pyreadline==2.1
pytest==6.2.3
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==22.3.0
QDarkStyle==2.8.1
QtPy==1.9.0
simplegeneric==0.8.1
sip==5.0.0
sqlparse==0.4.2
tables==3.6.1
unicodecsv==0.14.1
webencodings==0.5.1
whitenoise==5.3.0
widgetsnbextension==3.5.1
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.12.1
xlwt==1.3.0
zict==2.0.0
zope.event==4.5.0
jmetalpy==1.5.5